refactor(routes): move kalooki pages into kalooki feature module
